+++ mrbob/cli.py
@@ -47,6 +47,10 @@ parser.add_argument('-l', '--list-questions',
 parser.add_argument('-r', '--renderer',
                     action="store",
                     help='Dotted notation to a renderer function. Defaults to mrbob.rendering:jinja2_renderer')
+parser.add_argument('-q', '--quiet',
+                    action="store_true",
+                    default=False,
+                    help='Suppress all but necessary output')
 #parser.add_option('--simulate',
                   #dest='simulate',
                   #action='store_true',
@@ -138,7 +142,7 @@ def main(args=sys.argv[1:], quiet=False):
         if options.list_questions:
             return c.print_questions()
 
-        if c.questions:
+        if c.questions and not options.quiet:
             print("Welcome to mr.bob interactive mode. Before we generate directory structure, some questions need to be answered.")
             print("")
             print("Answer with a question mark to display help.")
@@ -147,7 +151,8 @@ def main(args=sys.argv[1:], quiet=False):
             c.ask_questions()
             print("")
         c.render()
-        print("Generated file structure at %s" % os.path.realpath(options.target_directory))
+        if not options.quiet:
+            print("Generated file structure at %s" % os.path.realpath(options.target_directory))
         return
     except TemplateConfigurationError as e:
         parser.error(six.u('TemplateConfigurationError: %s') % e.args[0])